+++ python/GafferSceneTest/__init__.py
@@ -160,6 +160,8 @@ from .NameSwitchTest import NameSwitchTest
 from .CryptomatteTest import CryptomatteTest
 from .ShaderQueryTest import ShaderQueryTest
 from .AttributeTweaksTest import AttributeTweaksTest
+from .OptionTweaksTest import OptionTweaksTest
+from .OptionQueryTest import OptionQueryTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *