+++ python/GafferSceneUI/__init__.py
@@ -174,6 +174,8 @@ from . import UVSamplerUI
 from . import CryptomatteUI
 from . import ShaderQueryUI
 from . import AttributeTweaksUI
+from . import OptionTweaksUI
+from . import OptionQueryUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.